Add image download buttons
In thumbnail view, the download button will appear in the lower right
corner of the image when hovering it.
In fullscreen view, the button is permanently displayed in the upper
right corner of the screen.
On very narrow devices (phones) the button is permanently visible in
thumbnail view and nonexistent in full screen view.


Former-commit-id: c2411e3f86b9f0a6df7d307cfeed5b8061d0c925