Update docker-compose.yml
Corrected volume path. Previously it mapped photos dir to photos/original, so the result was in the container a dir structure that goes photos/original/original. So the photos from the volume landed in photos/original/original and not in photos/originals. That is why there was an error saying that there are no images found.

Former-commit-id: 21a75a35d8f71556f07997f8e67cea715334eaf1